+++ src/tooling/piral-cli/src/types/public.ts
@@ -231,7 +231,7 @@ export type PiletPublishScheme = 'none' | 'digest' | 'bearer' | 'basic';
 
 export type PiletPublishSource = 'local' | 'npm' | 'remote';
 
-export type PiralBuildType = 'all' | 'release' | 'emulator' | 'emulator-sources';
+export type PiralBuildType = 'all' | 'release' | 'emulator' | 'emulator-sources' | 'emulator-website';
 
 export type PiletBuildType = 'default' | 'standalone' | 'manifest';